Moving function to provider file
